requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
watchdog==3.0.0
